Move controllers to packages; Move views to shared; Add template for configs
