+++ src/configs/templates/core.py
@@ -1,0 +1,13 @@
+template = {
+    "credentials": {},
+    "defaults": {
+        "package": "tapis",
+        "user": "",
+        "auth_method": "password"
+    },
+    "tapis.defaults": {
+        "tenant": "tacc",
+        "env": "prod",
+        "base_url": "https://tacc.tapis.io"
+    }
+}
