+++ src/core/TapipyController.py
@@ -1,4 +1,4 @@
-import options.handlers
+import shared.options.handlers
 
 from configs import settings
 from core.Controller import Controller
@@ -40,11 +40,11 @@ class TapipyController(Controller):
                 
                 # If the current option from the option set HAS been provided but there is
                 # no handler specified, ignore it
-                if option.handler == None or not hasattr(options.handlers, option.handler):
+                if option.handler == None or not hasattr(shared.options.handlers, option.handler):
                     continue
                 
                 # Register the handler
-                handlers[option.context].append(getattr(options.handlers, option.handler))
+                handlers[option.context].append(getattr(shared.options.handlers, option.handler))
             
             for handler in handlers["generic"]:
                 handler(self)
