+++ src/packages/tapis/controllers/Apps.py
@@ -3,7 +3,7 @@
 import json
 import sys
 
-from core.TapisController import TapisController
+from packages.tapis.TapisController import TapisController
 from tapipy.errors import InvalidInputError, ServerDownError
