+++ src/packages/tapis/controllers/Files.py
@@ -2,7 +2,7 @@
 
 import os
 
-from core.TapisController import TapisController
+from packages.tapis.TapisController import TapisController
 from tapipy.errors import InvalidInputError
