+++ src/packages/tapis/controllers/Jobs.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 import json
 
-from core.TapisController import TapisController
+from packages.tapis.TapisController import TapisController
 from tapipy.errors import InvalidInputError
