+++ src/packages/tapis/controllers/Tenants.py
@@ -1,6 +1,6 @@
 """Handles TAPIS functionality related to tenants."""
 
-from core.TapisController import TapisController
+from packages.tapis.TapisController import TapisController
 from tapipy.errors import InvalidInputError
